extract sip template tests to its own file
